requirements-doc.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
--find-links wheels

alabaster==0.7.12         # via sphinx
babel==2.5.3              # via E-Commerce, sphinx
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
django==1.11.20
django-bootstrap3==9.1.0  # via E-Commerce
django-countries==5.2     # via E-Commerce
django-enumfields==0.9.0  # via E-Commerce
django-filer==1.4.4       # via E-Commerce
django-filter==1.1.0      # via E-Commerce
django-jinja==2.4.1       # via E-Commerce
django-js-asset==1.2.2    # via django-mptt
django-mptt==0.9.1
django-parler==1.9.2      # via django-parler-rest, E-Commerce
django-parler-rest==1.4.2  # via E-Commerce
django-polymorphic==2.0.3  # via django-filer, E-Commerce
django-registration-redux==1.8  # via E-Commerce
django-rest-swagger==2.1.2  # via E-Commerce
django-reversion==2.0.13  # via E-Commerce
django-timezone-field==2.1  # via E-Commerce
djangorestframework==3.5.4  # via django-parler-rest, django-rest-swagger, E-Commerce
djangorestframework-jwt==1.11.0  # via E-Commerce
docutils==0.14            # via sphinx
easy-thumbnails==2.6      # via django-filer
entrypoints==0.3          # via keyring
et-xmlfile==1.0.1         # via openpyxl
factory-boy==2.7.0        # via E-Commerce
fake-factory==0.5.3       # via factory-boy, E-Commerce
idna==2.8                 # via requests
imagesize==1.1.0          # via sphinx
itypes==1.1.0             # via coreapi
jdcal==1.4                # via openpyxl
jinja2==2.8.1             # via coreschema, django-jinja, E-Commerce, sphinx
jsonfield==2.0.2          # via E-Commerce
keyring==19.0.1           # via E-Commerce
keyrings.alt==3.1.1       # via E-Commerce
lxml==4.2.1               # via E-Commerce
markdown==2.6.11          # via E-Commerce
markupsafe==1.1.1         # via jinja2
openapi-codec==1.3.2      # via django-rest-swagger
openpyxl==2.5.14          # via E-Commerce
pillow==5.4.1             # via easy-thumbnails
pygments==2.3.1           # via sphinx
pyjwt==1.7.1              # via djangorestframework-jwt
python-dateutil==2.8.0    # via E-Commerce
pytz==2018.9              # via babel, django, django-timezone-field
requests==2.21.0          # via coreapi, E-Commerce, sphinx
E-Commerce==1.9.2         # Shuup
simplejson==3.16.0        # via django-rest-swagger
six==1.12.0               # via keyrings-alt, python-dateutil, E-Commerce, sphinx
snowballstemmer==1.2.1    # via sphinx
sphinx==1.5.6
toml==0.10.0              # via E-Commerce
unicodecsv==0.14.1        # via E-Commerce
unidecode==1.0.23         # via django-filer
uritemplate==3.0.0        # via coreapi
urllib3==1.24.1           # via requests
xlrd==1.2.0               # via E-Commerce
